Fixing lint errors and renaming Pym file.
