+++ client/coral-framework/actions/asset.js
@@ -2,4 +2,4 @@ import * as actions from '../constants/asset';
 
 export const fetchAssetRequest = () => ({type: actions.FETCH_ASSET_REQUEST});
 export const fetchAssetSuccess = asset => ({type: actions.FETCH_ASSET_SUCCESS, asset});
-export const fetchAssetFailure = error => ({type: actions.FETCH_ASSET_FAILURE, error})
+export const fetchAssetFailure = error => ({type: actions.FETCH_ASSET_FAILURE, error});
